XW-2605 | move article media logic to a separate file
