+++ front/main/app/components/article-media-gallery.js
@@ -47,7 +47,7 @@ export default Ember.Component.extend(
 
 		actions: {
 			openLightbox(galleryRef) {
-				// openLightbox is set in getAttributesForMedia() inside components/article-content.js
+				// openLightbox is set in getAttributesForMedia() inside utils/article-media.js
 				this.get('openLightbox')(this.get('ref'), galleryRef);
 			}
 		},
