+++ front/main/app/components/article-media-thumbnail.js
@@ -45,7 +45,7 @@ export default Ember.Component.extend(
 
 		actions: {
 			openLightbox() {
-				// openLightbox is set in getAttributesForMedia() inside components/article-content.js
+				// openLightbox is set in getAttributesForMedia() inside utils/article-media.js
 				// it can also be overriden when this component is rendered from a template instead of JS
 				this.get('openLightbox')(this.get('ref'));
 			}
